request manager prints error
on a rrq, if the server recieved an error, it would not print the error message since it thought it would always recieve acks and had the length set to 4. Now it reads the packet and stores it in a tembuffer of size 1000 bytes. If it is an ack, it continues as normal, if it is of type error, then the server quits.

HOWEVER...need to find a way to make it so that if it's error type 5 that it doesn't exit.